[form-factors] Split analytic-b-to-p-lcsr.cc file and adapt Makefile.am
