+++ eos/form-factors/analytic-b-to-k-lcsr.cc
@@ -2,6 +2,7 @@
 
 /*
  * Copyright (c) 2018 Danny van Dyk
+ * Copyright (c) 2019 Nico Gubernari
  *
  * This file is part of the EOS project. EOS is free software;
  * you can redistribute it and/or modify it under the terms of the GNU General
@@ -23,23 +24,20 @@ namespace eos
 {
     namespace lcsr
     {
-        // B -> pi
-        constexpr const char * BToPi::P;
-        constexpr const char * BToPi::m_P;
-        constexpr const char * BToPi::f_P;
+        struct BToK
+        {
+            constexpr static const char * P    = "K";
+            constexpr static const char * m_P  = "mass::K_d";
+            constexpr static const char * f_P  = "decay-constant::K";
+            constexpr static const char   q_v  = 's';
+            constexpr static const double chi2 = 1.0;
+        };
 
         // B -> K
         constexpr const char * BToK::P;
         constexpr const char * BToK::m_P;
         constexpr const char * BToK::f_P;
-
-        // B -> D
-        constexpr const char * BToD::P;
-        constexpr const char * BToD::m_P;
-        constexpr const char * BToD::f_P;
     }
 
-    template class AnalyticFormFactorBToPLCSR<lcsr::BToPi>;
     template class AnalyticFormFactorBToPLCSR<lcsr::BToK>;
-    template class AnalyticFormFactorBToPLCSR<lcsr::BToD>;
 }
